Перенёс константы в отдельный файл в констаты)
